refactor: extract market detail logic into custom hook for MarketDetailV2
- Created a new custom hook, useMarketDetailData, to encapsulate the logic for fetching and managing market token details, improving code organization and reusability.
- Simplified the MarketDetailV2 component by utilizing the new hook, enhancing readability and maintainability of the codebase.
